Add screen_data utility
Add utility to get screen data.
This data is currently embedded as an application/json element
on the page and the screen_data utility fetches that data.
It's added to the entry point for the screen and embedded
as a property on the screen element.

Also, this patch adds "debug" as a utility for logging.
screen_data will log to "sycnhrotron:error" if it's asked to
find an HTML element that does not exist.