Add edit functionality to individual coupons.
Add react code to change the coupon display to an edit form.
Then add redux action code to track the edit state.
Also, affix package.json versions of React and Redux to match
Calypso because it is using hard versioning. This was causing
a problem within React in regards to setting ref attributes.