Add webpack dev server support.
This commit adds the dev server run from "npm start"
but there were some other changes necessary to make this work.
The wp_enqueue_script and wp_enqueue_style calls had to be
pointed to the webpack dev server, but only when it was running.
To do this, the webpack config was split, and relies on a new
php config file for each one to be copied in and included.

In addition, since the hashing for the URLs normally used
from the webpack dev server can't be used with enqueue, I also
implemented a random hex string for versioning that is configurable.

To test:

Run "npm run build-dev" and go to the Synchrotron coupons screen.
View source and search for "bundle". Verify that the URL given is
from the plugin's base URL and the version is a random hex string.

Then run "npm start" and go to the Synchrotron coupons screen.
View source and search for "bundle". Verify that the URL given is
from localhost:8090 and the version is a random hex string.