Move screenData call out of state/actions for tax-rates.
The screenData import was causing problems with running the tests
due to a lack of a window global, where the state code probably shouldn't
be depending on a browser global at that point. So this commit moves
the culling of that data to the top-level React container component for
tax rates.