Lock down wp-calypso to previous version
The latest wp-calypso updates are causing enough issues with this
branch that it's easier to lock down the version than it would be
to clutter this branch with what it would take to get current.

The strategy then is to get this branch merged back in to master
and update to a newer version of calypso in a separate patch.