Remove wp-calypso/node_modules/react on build.
Even though the version numbers match perfectly, npm insists
on downloading two copies of react. One at the top-level of
node_modules, and one under wp-calypso. This causes problems
on build with React.

`npm dedupe` fixes this, but takes a REALLY long time. So I've
added it as a build step to speed thing along and make sure new
folks don't have to face this issue.