Try out some simple popover operation code.
After much deliberation of how to best accomplish this feat,
I've settled on getting the context for the "anchoring" component
for the Popover from the actual event itself. This  makes it easy for
a single component to support a "hover" popover (i.e. tooltip) as well
as a larger "click" popover menu. It would also allow a single
Popover to service multiple anchoring components.