Add constraints parameters to columns.
Instead of the different render functions having optional parameters at the end,
these are now consolidated into a constraints object that belongs to each column
object and passed to the render function. This cleans up and organizes the little
bits of extra data that each render function might care about.