Add fetch-data state cache
This adds a state cache for commonly fetched data.
Examples: categories, tags, products, etc.

This allows the fetched data to be stored and used by other parts
of the application when needed. This will save round-trip HTTP requests
by various parts of the code since we'll already have this data in memory.