Product List: Make column groups their own entities
This commit makes column groups their own first-class objects.
This starts to remove the 1:1 coupling between columns and selections
from the menu. A nice side effect of this is less manipulating of the
array for pulling out the groups, and also it allows a column to be
first in the menu, but last in the display order.