product state: fix tests
The referencing of columns.js from the products reducer introduces
JSX into the reducer test environment, which breaks testing because
of the gridicon dependency, which has a different root than the client.

This also isn't exactly right anyway, because the reducer shouldn't be
dependent on columns.js as it's an application-specific file and the
reducer is a component-specific file that should have no knowledge of
the application-specific definitions.

This, therefore fixes it for now by removing this dependency.
However, it does uncover a looming issue for testing JSX that might
contain a gridicon or wp-calypso component. That issue is logged
here: https://github.com/Automattic/wc-synchrotron/issues/72