WIP fetch-data: Consolidate fetch data sets.
This commit starts to consolidate each fetch data set into an object
each with an action and data function. The action function will be used
to fetch the data when needed, and the data function is used to retrieve
the data from the appropriate place in the fetch data state.