WIP fetch-data: Use HOC for fetch-data
This commit creates the `fetchConnect()` function which is a
Higher Order Component that piggybacks on the `redux-react`
`connect()` HOC. It adds the function `getFetchProps()` which
provides a list of props as arguments. These props from the component
can be used to construct queries for the fetch data for things such
as search queries, pagination, etc.

Note: this commit only handles the display of the data, the actual
fetching is still done like it was before, hence the WIP.