fetch-data: Make fetches more declarative.
Translate all of the fetch info into declarative data wherever possible.
(This excludes only the action to do the fetch)
This will allow the fetch-data logic to know more about the nature of
the fetch data, like the service, key, and default value.
It will also make it simpler for new fetches to be added.