fetch-data: FetchConnct cache to fetch props data.
This commit holds references to the fetch props data. This allows
us to determine if the fetch props data has changed when the redux
store state changes. If our own props change, or our fetch props data
changes, then we signal via `shouldComponentUpdate` to let React
know that we need to be re-rendered.

This commit also fixes a couple minor code issues that were found
with the unit tests coming in the next commit.