fetch-data: Add chai tests for FetchConnect
The FetchConnect HOC did not have any tests for it, so this
commit adds some. They're not 100% comprehensive as we would
have to load up JSDom or run these in the browser to accomplish
that. But using `skin-deep`, it was possible to test almost all
normal functionality.

Note: Error tests are not included yet, as that functionality is
still to be done.