fetch-data: Map state objects instead of data.
The fetch data has been moved down to a `data` property
of a `fetchState` object. This object will also hold error
messages and other metadata that would be useful to the component.

So, this change now requires components to use `myProp.data` to
access the fetched data. But it will also allow components to use
`myProp.error` to access any error info for that fetch as well.