fetch-data: Move error and metadata
In the last PR, the error status and metadata were added to the fetch
data structure. But considering this is really metadata, it makes sense
to place it in its own area in the store. Therefore this commit moves
that metadata to its own location under fetch-data in the store, to keep
it separate from the components, which will rarely ever care about such
state.