fetch-data: use { value, status } structure
For the fetch data, I've consolidated on the structure of:

{ value, status }

This is better than the previous one of sibling objects where one was
the key and the other is key + '_status' for two reasons:

1. key + '_status' could be spoofed by another key.
2. This keeps all data concerning a given fetch under a single object.

The second point is important when we start expiring fetches and we want
to reclaim that memory. This would allow us to do that with one object
modification instead of two.