fetch-data: Code to clean up expired fetches.
This adds the code required to clean up the expired fetch metadata and
the ability to send an action to trigger the deletion of the fetch data
from the redux store.

This commit does not include the cleaning up of that data, nor a timer
setup to run the cleanup task, but it does include the cleanup function
itself and a test for it.